# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScreenCrew API - friend-group screen sharing app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, isAdmin, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        steamUrl:
          type: string
          nullable: true
        discordUrl:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        nameColor:
          type: string
          nullable: true
        avatarStyle:
          type: string
          nullable: true
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    UpdateProfileInput:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
          maxLength: 40
        email:
          type: string
          nullable: true
        steamUrl:
          type: string
          nullable: true
        discordUrl:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        nameColor:
          type: string
          nullable: true
          maxLength: 9
        avatarStyle:
          type: string
          nullable: true
          enum: [initials, pixel]
...